Add missing unit tests for node status updates
- Extracting out status related functions in a separate file
